[Backport] Security bug 905509 (12/13)
Backport of original patch by  Antoine Labour <piman@chromium.org>:
Sanitize use of size_t in {Service,Client}FontManager and
RasterCHROMIUM

Because transfer buffers are 4GB at the max, use uint32_t instead of
size_t (or uint64_t) for sizes in font manager serialized data.

Bug: 905509
Reviewed-on: https://chromium-review.googlesource.com/c/1404390

+++ chromium/gpu/command_buffer/client/raster_implementation.h
@@ -167,7 +167,7 @@ class RASTER_EXPORT RasterImplementation : public RasterInterface,
   void UnmapRasterCHROMIUM(uint32_t written_size);
 
   // ClientFontManager::Client implementation.
-  void* MapFontBuffer(size_t size) override;
+  void* MapFontBuffer(uint32_t size) override;
 
  private:
   friend class RasterImplementationTest;
